requirements.txt
altgraph==0.17
dnspython==2.0.0
numpy==1.19.4
paho-mqtt==1.5.1
pkg-resources==0.0.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.10
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyserial==3.4
python-etcd==0.4.5
urllib3==1.26.2
